Version 2.2.0 - Updated SNP insert logic
Previously, duplicate strainid/locus will cause an exception.  This has been
handled so that if the base is the same (i.e. a true duplicate), then it is
ignored.  If the bases are DIFFERENT (so the same strain id and locus has
different reported bases in the snp data) then it is a hard error.